requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
cryptography==45.0.5
exceptiongroup==1.1.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.4
iniconfig==2.0.0
kaitaistruct==0.10
outcome==1.2.0
packaging==23.1
pluggy==1.2.0
pyasn1==0.6.1
pycparser==2.22
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
selenium-page-factory==2.7
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.14.1
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
zstandard==0.23.0
